rewrote and moved buildSearchSettingsFromStoredProjects to @Common. Removed rootWeb from projectstats
