+++ src/js/@WebParts/ProgramDeliveriesOverview/index.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 import IProgramDeliveriesOverviewProps, { ProgramDeliveriesOverviewDefaultProps } from "./IProgramDeliveriesOverviewProps";
 import IProgramDeliveriesOverviewState, { } from "./IProgramDeliveriesOverviewState";
 import DeliveriesOverview from "prosjektportalen/lib/WebParts/DeliveriesOverview";
-import { IDataSourceSearchCustom } from "prosjektportalen/lib/WebParts/DataSource";
 import * as common from "../../@Common";
 import DataSource from "prosjektportalen/lib/WebParts/DataSource";
 import { MessageBar, MessageBarType } from "office-ui-fabric-react/lib/MessageBar";
@@ -19,7 +18,8 @@ export default class ProgramDeliveriesOverView extends React.Component<IProgramD
 
   public async componentDidMount() {
     try {
-      const searchSettings = await this.buildSearchSettingsFromStoredProjects();
+      const { items } = await common.getStoredProjectsListContext();
+      const searchSettings = await common.buildSearchSettingsFromStoredProjects(items, this.props.queryTemplate);
       this.setState({ searchSettings, isLoading: false });
     } catch (errorMessage) {
       this.setState({ errorMessage, isLoading: false });
@@ -52,25 +52,6 @@ export default class ProgramDeliveriesOverView extends React.Component<IProgramD
             queryTemplate={this.state.searchSettings.QueryTemplate} />}
       </>
     );
-  }
-
-  /**
-   * Build search settings from items in stored projects list
-   */
-  private async buildSearchSettingsFromStoredProjects(): Promise<IDataSourceSearchCustom> {
-    try {
-      const { items } = await common.getStoredProjectsListContext();
-      if (items.length === 0) {
-        return null;
-      }
-      const searchQuery = items.map(({ URL }) => `Path:"${URL}"`).join(" OR ");
-      return {
-        RowLimit: 500,
-        QueryTemplate: String.format(this.props.queryTemplate, searchQuery),
-      };
-    } catch (err) {
-      throw err;
-    }
   }
 
 }
