+++ src/js/@WebParts/ProgramProjectStats/IProgramProjectStatsState.ts
@@ -1,11 +1,9 @@
 import { ProjectItem } from "src/js/@Common";
-import { Web } from "@pnp/sp";
 import { IDataSourceSearchCustom } from "prosjektportalen/lib/WebParts/DataSource";
 
 export interface IProgramProjectStatsState {
     isLoading: boolean;
     errorMessage?: string;
-    rootWeb?: Web;
     items?: ProjectItem[];
     searchSettings?: IDataSourceSearchCustom;
 }
