+++ src/js/@WebParts/ProgramResourceAllocation/index.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 import IProgramResourceAllocationProps, { ProgramResourceAllocationDefaultProps } from "./IProgramResourceAllocationProps";
 import { IProgramResourceAllocationState } from "./IProgramResourceAllocationState";
 import ResourceAllocation from "prosjektportalen/lib/WebParts/ResourceAllocation";
-import { IDataSourceSearchCustom } from "prosjektportalen/lib/WebParts/DataSource";
 import * as common from "../../@Common";
 import DataSource from "prosjektportalen/lib/WebParts/DataSource";
 import { MessageBar, MessageBarType } from "office-ui-fabric-react/lib/MessageBar";
@@ -20,7 +19,7 @@ export default class ProgramResourceAllocation extends React.Component<IProgramR
   public async componentDidMount() {
     try {
       const { items } = await common.getStoredProjectsListContext();
-      const searchSettings = await this.buildSearchSettingsFromStoredProjects(items);
+      const searchSettings = await common.buildSearchSettingsFromStoredProjects(items, this.props.queryTemplate);
       this.setState({ searchSettings, isLoading: false });
     } catch (errorMessage) {
       this.setState({ errorMessage, isLoading: false });
@@ -55,24 +54,6 @@ export default class ProgramResourceAllocation extends React.Component<IProgramR
           />}
       </>
     );
-  }
-
-  /**
-   * Build search settings from items in stored projects list
-   */
-  private async buildSearchSettingsFromStoredProjects(items: common.ProjectItem[]): Promise<IDataSourceSearchCustom> {
-    try {
-      if (items.length === 0) {
-        return null;
-      }
-      const searchQuery = items.map(({ URL }) => `Path:"${URL}"`).join(" OR ");
-      return {
-        RowLimit: 500,
-        QueryTemplate: String.format(this.props.queryTemplate, searchQuery),
-      };
-    } catch (err) {
-      throw err;
-    }
   }
 
 }
